// src/db/generated/models/AcceptedIssuer.ts
export type AcceptedIssuerModel = runtime.Types.Result.DefaultSelection<Prisma.$AcceptedIssuerPayload>
// ...
export type $AcceptedIssuerPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "AcceptedIssuer"
  objects: {
    agency: Prisma.$AgencyPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    name: string
    kind: $Enums.AcceptedIssuerKind
    retiredAt: Date | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["acceptedIssuer"]>
  composites: {}
}